backend/requirements.txt
asgiref==3.10.0
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
ConfigArgParse==1.7.1
Django==5.2.7
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.41.6
locust-cloud==1.27.6
MarkupSafe==3.0.3
msgpack==1.1.2
mysqlclient==2.2.7
packaging==25.0
platformdirs==4.5.0
pluggy==1.6.0
psutil==7.1.0
Pygments==2.19.2
pytest==8.4.2
python-engineio==4.12.3
python-socketio==5.14.2
pyzmq==27.1.0
redis==6.4.0
requests==2.32.5
setuptools==80.9.0
simple-websocket==1.1.0
sqlparse==0.5.3
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==6.0
zope.interface==8.0.1
